app/requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
DateTime==4.3
Flask==1.1.2
Flask-JWT-Extended==4.1.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.2
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
Werkzeug==1.0.1
zope.interface==5.3.0
cryptography==2.8
